Adding calibration routines and restructuring modules a bit
